set protein dict to have arrays of len 1 always
when creating a pandas dataframe the arrays must be equal lengths. The lists within the links dictionary could be different lengths and would thus raise an error when creating the protein dataframe. All links within the links database are written to a string, per database, so the length of the array for each database is 1