trim README.md significantly
Major edits to README.md including removal of redundant and outdated
information; moving sections of text to the repository wiki; tidying
and clarification of text.

The README.md file should be short and introduce a new user quickly
to the main concepts of the package/tool. If tool usage is simple
and/or there is no additional documentation, then the README.md is
a good place for usage and other information. In this case, there
is information on ReadTheDocs and in the repository wiki, so we
do not need to keep redundand information here.

In general, we don't want to store the same information in multiple
places (e.g. ReadTheDocs, wiki, README.md) because we need to
monitor and update several different locations for a single change.
That is time-consuming and wasteful. It is inevitable that documents
will start to contradict each other, and cause confusion.