rationalise installation instructions
The main changes are:

- remove unnecessary tutorials on using the shell, etc.

We cannot cover all circumstances that a novice user might
encounter. Providing instructions for software we do not maintain
or control should be considered out of scope. It is fair to
point to other tutorials, but we should restrain ourselves as
much as possible to concise instructions for our own tools and
applications.

- more precise instructions for local installation