create make_target_dir and make_output_dir
make_target_dir() creates the target dir for the output file path passed to the function, it does not worry if there is already contents in the output dir, it will write to it anyway. make_output_dir() creates a dir the dir path passed to it, this dir will have multiple output files written to it and cares if the output dir already exists and contains output files. The nodelete param determines if the existing output in the output dir is deleted or not