add checking if family is empty when 404 err raised
Empty and deleted families raise an HTTP 404 error when trying to download the family txt file, this is expected and accepted. If the family is not empty or deleted then this is logged. An attempt to still retrieve the family text file for empty and delted families incase there is data in it and the family summary table on the website is incorrect