simplifying getting assembly data
do not fetch nuccore records and extract data. Link proteins to nuccore, get nuccore ids, link nuccore to assembly, get assembly ids and fetch assembly records, then link proteins to assemblies when download assemlies. At that point then identify which proteins no assembly was retrieved. It means it won't log when the issue was a missing protein-nuccore link or nuccore-assembly link but it will get the overally missing protein-assembly link, will be faster and less demanding on the Entrez server