use the uniprot mapping service
Use the UniProt mapping service to map the uniprot accession to the genbank protein version accession. For the proteins for which that fails, if could not retrieve a gene name from the original UniProt query, map the uniprot accession to the gene name. For proteins for which they could not map the uniprot acc to genbank accession, query ncbi by the gene name to retrieve the genbank accession